+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **[nitro]** Add `defineConfig` helper function.
+- **[cli]** Export `defineConfig` helper function directly from `yummacss` package.
 
 ## [3.18.0] - 2026-03-04